fix: use the targetBranch option to reset/pr
Use the targetBranch option to reset the working directory to the correct git
ref and open a pull request on GitHub with the correct base branch. This is
usually the default branch like main but could be any pre-existing branch on
GitHub. At the moment, the targetBranch option is exposed but it's no used.

**Note**: This change keeps the existing `master` default value but we'd like
to change that default to `main` along with a breaking version bump.